+++ src/main/java/com/microsoft/azure/functions/worker/broker/FunctionExecutionPayLoad.java
@@ -6,12 +6,10 @@ import com.microsoft.azure.functions.worker.description.FunctionMethodDescriptor
 import com.microsoft.azure.functions.worker.reflect.ClassLoaderProvider;
 
 import java.lang.reflect.Method;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
 public class FunctionExecutionPayLoad {
     private final Class<?> containingClass;
-    private final ParameterResolver overloadResolver;
     private final Map<String, BindingDefinition> bindingDefinitions;
 
     public FunctionExecutionPayLoad(FunctionMethodDescriptor descriptor, Map<String, BindingInfo> bindingInfos, ClassLoaderProvider classLoaderProvider)
@@ -19,20 +17,20 @@ public class FunctionExecutionPayLoad {
     {
         descriptor.validateMethodInfo();
 
+        this.candidates = new ArrayList<>();
         this.containingClass = getContainingClass(descriptor.getFullClassName(), classLoaderProvider);
-        this.overloadResolver = new ParameterResolver();
 
         for (Method method : this.containingClass.getMethods()) {
             if (method.getDeclaringClass().getName().equals(descriptor.getFullClassName()) && method.getName().equals(descriptor.getMethodName())) {
-                this.overloadResolver.addCandidate(method);
+                this.addCandidate(method);
             }
         }
 
-        if (!this.overloadResolver.hasCandidates()) {
+        if (!hasCandidates()) {
             throw new NoSuchMethodException("There are no methods named \"" + descriptor.getName() + "\" in class \"" + descriptor.getFullClassName() + "\"");
         }
 
-        if (this.overloadResolver.hasMultipleCandidates()) {
+        if (hasMultipleCandidates()) {
             throw new UnsupportedOperationException("Found more than one function with method name \"" + descriptor.getName() + "\" in class \"" + descriptor.getFullClassName() + "\"");
         }
 
@@ -52,11 +50,25 @@ public class FunctionExecutionPayLoad {
         return containingClass;
     }
 
-    public ParameterResolver getOverloadResolver() {
-        return overloadResolver;
-    }
-
     public Map<String, BindingDefinition> getBindingDefinitions() {
         return bindingDefinitions;
     }
+
+
+    synchronized void addCandidate(Method method) {
+        this.candidates.add(new MethodBindInfo(method));
+    }
+
+    public synchronized boolean hasCandidates() {
+        return !this.candidates.isEmpty();
+    }
+
+    public synchronized boolean hasMultipleCandidates() {
+        return this.candidates.size() > 1;
+    }
+
+    public synchronized MethodBindInfo getMethodBindInfo() {
+        return this.candidates.get(0);
+    }
+    private final List<MethodBindInfo> candidates;
 }
