+++ src/main/java/com/microsoft/azure/functions/worker/broker/InvokeInfoBuilder.java
@@ -1,0 +1,12 @@
+package com.microsoft.azure.functions.worker.broker;
+
+import java.util.UUID;
+
+public final class InvokeInfoBuilder extends JavaMethodInvokeInfo.Builder {
+    public InvokeInfoBuilder(MethodBindInfo method) { super.setMethod(method.getEntry()); }
+    private final UUID outputsId = UUID.randomUUID();
+
+    public UUID getOutputsId() {
+        return outputsId;
+    }
+}
