--- src/main/java/com/microsoft/azure/functions/worker/broker/ParameterResolver.java
@@ -1,54 +1,0 @@
-package com.microsoft.azure.functions.worker.broker;
-
-import java.lang.invoke.WrongMethodTypeException;
-import java.lang.reflect.Method;
-import java.lang.reflect.Parameter;
-import java.lang.reflect.Type;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-import org.apache.commons.lang3.exception.ExceptionUtils;
-import org.apache.commons.lang3.reflect.TypeUtils;
-
-import com.microsoft.azure.functions.OutputBinding;
-import com.microsoft.azure.functions.worker.binding.BindingData;
-import com.microsoft.azure.functions.worker.binding.BindingDataStore;
-
-/**
- * Resolve a Java method overload using reflection.
- * Thread-Safety: Multiple thread.
- */
-public class ParameterResolver {
-    ParameterResolver() {
-        this.candidates = new ArrayList<>();
-    }
-
-    synchronized void addCandidate(Method method) {
-        this.candidates.add(new MethodBindInfo(method));
-    }
-
-    public synchronized boolean hasCandidates() {
-        return !this.candidates.isEmpty();
-    }
-
-    public synchronized boolean hasMultipleCandidates() {
-        return this.candidates.size() > 1;
-    }
-
-    public synchronized MethodBindInfo getMethodBindInfo() {
-        return this.candidates.get(0);
-    }
-  
-    public static final class InvokeInfoBuilder extends JavaMethodInvokeInfo.Builder {
-        public InvokeInfoBuilder(MethodBindInfo method) { super.setMethod(method.getEntry()); }
-        private final UUID outputsId = UUID.randomUUID();
-
-        public UUID getOutputsId() {
-            return outputsId;
-        }
-    }
-    private final List<MethodBindInfo> candidates;
-}
